Add note about scheduling format - removed form.
I used to have a PHP-backed contact form that used PHP Mailer. I didn't want to
pay for hosting and most contact with students was via text message anyway,
though, so I'm hosting the site on GitHub pages without a back end and asking
students to text, call, or email to schedule an appointment.
Will it work as well? I'll find out.