Change to single "midland" branch.
I kept the desktop and mobile sites in sync with regard to content. (The
process was inefficient but they were in sync.) Therefore, I don't see a
reason to have separate branches for desktop and mobile.

Desktop stuff is now in the desktop directory; mobile in mobile - same as
before I had the site under version control.